repl: Enables history for programmatic REPLs.
The commit moves the `setupHistory()` function from `internal/repl.js`
to the public facing `REPLServer` API. Default behavior has not changed
for either the command line or the programmatic REPL. The
`createInternalRepl` function used by `src/node.js` now simply calls
`setupHistory()` on the `REPLServer` it creates.

History size for the programmatic REPL may be controlled by providing a
value for the heretofore undocumented `historySize` in the options for
`repl.start()`.
